requirements.txt
pandas
numpy
jupyter
python-dotenv